+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.2.0 (2021-08-23)
+* [bugfix] Fixed channel for tessa assets in entity records (when fetching them via api)
+* [feature] Added api endpoint for communication between Tessa and Akeneo 
+
 ## 1.1.0 (2021-08-11)
 * [feature] Adjust the amount of assets displayed in product grid and reference entity grid
 * [feature] Added asset download links in API